% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contingencyTables.R
\name{contingencyTables}
\alias{contingencyTables}
\title{Contingency Tables}
\usage{
contingencyTables(x, y, data)
}
\arguments{
\item{x}{A variable name given as a character string}

\item{y}{A variable name given as a character string}

\item{data}{The name of the dataframe.}
}
\value{
The contingency table of variables x and y
}
\description{
This function creates a cotingency table based on
two variables of a dataframe.
}
\examples{
data(dataSurveyA3)
contingencyTables("genero","municipio", dataSurveyA3)
contingencyTables("genero","municipio", dataSurveyA3)
}
