% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyA3Madrid.R
\docType{data}
\name{dataSurveyA3}
\alias{dataSurveyA3}
\title{Travel survey and stated preferences.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1186 rows and 72 columns.}
\usage{
data(dataSurveyA3)
}
\description{
A dataset containing information about a travel survey data carried out in September-October 2018
in two municipalities of the A3 Eastern Corridor Madrid: Rivas and Arganda.
}
\examples{
 data(dataSurveyA3)
 genero <- dataSurveyA3$genero
}
\keyword{datasets}
