% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyA3Madrid.R
\docType{package}
\name{surveyA3Madrid}
\alias{surveyA3Madrid}
\alias{surveyA3Madrid-package}
\title{surveyA3Madrid: A package with travel survey data from A3 Eastern Corridor, Madrid, Spain.}
\description{
This package contains information about a travel survey data carried out in September-October 2018
in two municipalities of the A3 Eastern Corridor Madrid: Rivas and Arganda. The package includes a sample function to create
contingency tables in a simple way.
}
\references{
\url{https://github.com/Fernan-tech}
}
\author{
Fernando Romero, TRANSyT Transport Research Centre, Universidad Politecnica de Madrid \email{fernando.romero@upm.es}
}
